Update heretic-host.ks
- Added support for LABEL= in inst.ks for parameters file retrieval
- Added support for OEMDRV labeled drives for parameters file retrieval
- Added safeguard for mounted disks and kickstart-source disks in disks initialization during %pre phase
- Allowed unspecified kickstart file pathname for parameters file retrieval
- Disabled use of CR repo
- Added safeguards for oVirt nightly repos use
- Changed CentOS-hosted oVirt repos from disabled to skip_if_unavailable when in nightly mode
- Small comment and code-style-uniformity fixes

+++ heretic-host.ks
@@ -270,46 +270,6 @@ ipmat() {
 	return 0
 }
 
-# Create install source selection fragment
-# Note: we use a non-local (hd:) stage2 location as indicator of network boot
-given_stage2=$(sed -n -e 's/^.*inst\.stage2=\(\S*\).*$/\1/p' /proc/cmdline)
-if echo "${given_stage2}" | grep -q '^hd:' ; then
-	# Detect use of NetInstall media
-	if [ -d /run/install/repo/repodata ]; then
-		# Note: we know that the local stage2 comes from a full DVD image (Packages repo included)
-		cat <<- EOF > /tmp/full-installsource
-		# Use the inserted optical media as in:
-		cdrom
-		# alternatively specify a NFS network share as in:
-		# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
-		# or an HTTP/FTP area as in:
-		#url --url https://dangerous.ovirt.life/hvp-repos/el7/os
-		EOF
-	else
-		# Note: since we detected use of NetInstall media (no local repo) we use network install source from CentOS mirrors
-		given_stage2="http://mirror.centos.org/centos/7/os/x86_64"
-		cat <<- EOF > /tmp/full-installsource
-		# Specify a NFS network share as in:
-		# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
-		# or an HTTP/FTP area as in:
-		url --url ${given_stage2}
-		# alternatively use the inserted optical media as in:
-		#cdrom
-		EOF
-	fi
-else
-	# Note: we assume that a remote stage2 has been copied together with the full media content preserving the default DVD structure
-	# TODO: we assume a HTTP/FTP area - add support for NFS
-	cat <<- EOF > /tmp/full-installsource
-	# Specify a NFS network share as in:
-	# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
-	# or an HTTP/FTP area as in:
-	url --url ${given_stage2}
-	# alternatively use the inserted optical media as in:
-	#cdrom
-	EOF
-fi
-
 # Define all cluster default network data
 # Note: engine-related data will only be used for automatic DNS zones configuration
 unset nicmacfix
@@ -492,13 +452,16 @@ notification_receiver="monitoring@localhost"
 ovirt_version="4.1"
 
 # Detect any configuration fragments and load them into the pre environment
-# Note: BIOS based devices, file and DHCP methods are unsupported
+# Note: incomplete (no device or filename), BIOS based devices, UUID, file and DHCP methods are unsupported
+ks_custom_frags="hvp_parameters.sh hvp_parameters_heretic_ngn.sh hvp_parameters_heretic_host.sh"
 mkdir /tmp/kscfg-pre
 mkdir /tmp/kscfg-pre/mnt
-ks_source="$(cat /proc/cmdline | sed -e 's/^.*\s*inst\.ks=\(\S*\)\s*.*$/\1/')"
+ks_source="$(cat /proc/cmdline | sed -n -e 's/^.*\s*inst\.ks=\(\S*\)\s*.*$/\1/')"
 if [ -z "${ks_source}" ]; then
-	echo "Unable to determine Kickstart source - skipping configuration fragments retrieval" 1>&2
-else
+	# Note: if we are here and no Kickstart has been explicitly specified, then it must have been found by OEMDRV method (needs CentOS >= 7.2)
+	ks_source='hd:LABEL=OEMDRV'
+fi
+if [ -n "${ks_source}" ]; then
 	ks_dev=""
 	if echo "${ks_source}" | grep -q '^floppy' ; then
 		# Note: hardcoded device name for floppy disk
@@ -510,7 +473,7 @@ else
 		if [ -z "${ks_path}" ]; then
 			ks_path="/ks.cfg"
 		fi
-		ks_dir="$(echo ${ks_path} | sed 's%/[^/]*$%%')"
+		ks_dir="$(echo ${ks_path} | sed -e 's%/[^/]*$%%')"
 	elif echo "${ks_source}" | grep -q '^cdrom:' ; then
 		# Note: cdrom gets accessed as real device name which must be detected - assuming it is the first removable device
 		# Note: hardcoded possible device names for CD/DVD - should cover all reasonable cases
@@ -525,10 +488,10 @@ else
 					ks_fsopt="ro"
 					ks_path="$(echo ${ks_source} | awk -F: '{print $2}')"
 					if [ -z "${ks_path}" ]; then
-						echo "Unable to determine Kickstart source path" 1>&2
-						ks_dev=""
+						ks_path="/ks.cfg"
+						ks_dir="/"
 					else
-						ks_dir="$(echo ${ks_path} | sed 's%/[^/]*$%%')"
+						ks_dir="$(echo ${ks_path} | sed -e 's%/[^/]*$%%')"
 					fi
 					break
 				fi
@@ -536,34 +499,46 @@ else
 		done
 	elif echo "${ks_source}" | grep -q '^hd:' ; then
 		# Note: blindly extracting device name from Kickstart commandline
-		ks_dev="/dev/$(echo ${ks_source} | awk -F: '{print $2}')"
+		ks_spec="$(echo ${ks_source} | awk -F: '{print $2}')"
+		ks_dev="/dev/${ks_spec}"
+		# Detect LABEL-based device selection
+		if echo "${ks_spec}" | grep -q '^LABEL=' ; then
+			ks_label="$(echo ${ks_spec} | awk -F= '{print $2}')"
+			if [ -z "${ks_label}" ]; then
+				echo "Invalid definition of Kickstart labeled device" 1>&2
+				ks_dev=""
+			else
+				ks_dev=/dev/$(lsblk -r -n -o name,label | awk "/\\<$(echo ${ks_label} | sed -e 's%\([./*\\]\)%\\\1%g')\\>/ {print \$1}" | head -1)
+			fi
+		fi
 		# TODO: Detect actual filesystem type on local drive - assuming VFAT
 		ks_fstype="vfat"
 		ks_fsopt="ro"
 		ks_path="$(echo ${ks_source} | awk -F: '{print $3}')"
 		if [ -z "${ks_path}" ]; then
-			echo "Unable to determine Kickstart source path" 1>&2
-			ks_dev=""
+			ks_path="/ks.cfg"
+			ks_dir="/"
 		else
-			ks_dir="$(echo ${ks_path} | sed 's%/[^/]*$%%')"
+			ks_dir="$(echo ${ks_path} | sed -e 's%/[^/]*$%%')"
 		fi
 	elif echo "${ks_source}" | grep -q '^nfs:' ; then
 		# Note: blindly extracting NFS server from Kickstart commandline
+		# TODO: support NFS options
 		ks_host="$(echo ${ks_source} | awk -F: '{print $2}')"
 		ks_fstype="nfs"
 		ks_fsopt="ro,nolock"
 		ks_path="$(echo ${ks_source} | awk -F: '{print $3}')"
 		if [ -z "${ks_path}" ]; then
-			echo "Unable to determine Kickstart source path" 1>&2
+			echo "Unable to determine Kickstart NFS source path" 1>&2
 			ks_dev=""
 		else
-			ks_dev="${ks_host}:$(echo ${ks_path} | sed 's%/[^/]*$%%')}"
+			ks_dev="${ks_host}:$(echo ${ks_path} | sed -e 's%/[^/]*$%%')}"
 			ks_dir="/"
 		fi
 	elif echo "${ks_source}" | egrep -q '^(http|https|ftp):' ; then
 		# Note: blindly extracting URL from Kickstart commandline
 		ks_host="$(echo ${ks_source} | sed -e 's%^.*//%%' -e 's%/.*$%%')"
-		ks_dev="$(echo ${ks_source} | sed 's%/[^/]*$%%')"
+		ks_dev="$(echo ${ks_source} | sed -e 's%/[^/]*$%%')"
 		ks_fstype="url"
 	else
 		echo "Unsupported Kickstart source detected" 1>&2
@@ -571,7 +546,6 @@ else
 	if [ -z "${ks_dev}" ]; then
 		echo "Unable to extract Kickstart source - skipping configuration fragments retrieval" 1>&2
 	else
-		ks_custom_frags="hvp_parameters.sh hvp_parameters_heretic_ngn.sh hvp_parameters_heretic_host.sh"
 		# Note: for network-based kickstart retrieval methods we extract the relevant nic MAC address to get the machine-specific fragment
 		if [ "${ks_fstype}" = "url" -o "${ks_fstype}" = "nfs" ]; then
 			# Note: we detect the nic device name as the one detaining the route towards the host holding the kickstart script
@@ -630,7 +604,7 @@ if grep -w -q 'hvp_nicmacfix' /proc/cmdline ; then
 	nicmacfix="true"
 fi
 
-# Determine choice of skipping custom packages intallation
+# Determine choice of skipping custom packages installation
 if grep -w -q 'hvp_orthodox' /proc/cmdline ; then
 	orthodox_mode="true"
 fi
@@ -656,14 +630,17 @@ fi
 if [ -b "/dev/${given_nodeosdisk}" ]; then
 	# If the given string is a device name then use that
 	disk_device_name="${given_nodeosdisk}"
+	disk_device_size=$(blockdev --getsize64 /dev/${disk_device_name})
 else
 	# If the given string is a generic indication then find the proper device
 	case "${given_nodeosdisk}" in
 		first)
 			disk_device_name=$(echo "${all_devices}" | head -1)
+			disk_device_size=$(blockdev --getsize64 /dev/${disk_device_name})
 			;;
 		last)
 			disk_device_name=$(echo "${all_devices}" | tail -1)
+			disk_device_size=$(blockdev --getsize64 /dev/${disk_device_name})
 			;;
 		*)
 			# Note: we allow for choosing either the first smallest device (default, if only "smallest" has been indicated) or the last one
@@ -738,7 +715,6 @@ for name in $(echo "${given_names}" | sed -e 's/,/ /g'); do
 		node_index=$((node_index+1))
 	fi
 done
-my_name="${node_name[${my_index}]}"
 
 # Determine installer name
 given_installer_name=$(sed -n -e 's/^.*hvp_installername=\(\S*\).*$/\1/p' /proc/cmdline)
@@ -830,6 +806,23 @@ if echo "${given_storage_offset}" | grep -q '^[[:digit:]]\+$' ; then
 	storage_ip_offset="${given_storage_offset}"
 fi
 
+# Determine hostname
+my_name="${node_name[${my_index}]}"
+
+# Determine AD subdomain name
+given_ad_subdomainname=$(sed -n -e "s/^.*hvp_ad_subdomainname=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
+if [ -n "${given_ad_subdomainname}" ]; then
+	ad_subdomain_prefix="${given_ad_subdomainname}"
+fi
+
+# Determine Gluster volume names
+for volume in engine vmstore iso ctdb unixshare winshare blockshare backup; do
+	given_volume_name=$(sed -n -e "s/^.*hvp_${volume}_volumename=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
+	if [ -n "${given_volume_name}" ]; then
+		gluster_vol_name["${volume}"]="${given_volume_name}"
+	fi
+done
+
 # Determine nameserver address
 given_nameserver=$(sed -n -e "s/^.*hvp_nameserver=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
 if [ -n "${given_nameserver}" ]; then
@@ -847,20 +840,6 @@ given_gateway=$(sed -n -e "s/^.*hvp_gateway=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
 if [ -n "${given_gateway}" ]; then
 	my_gateway="${given_gateway}"
 fi
-
-# Determine AD subdomain name
-given_ad_subdomainname=$(sed -n -e "s/^.*hvp_ad_subdomainname=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
-if [ -n "${given_ad_subdomainname}" ]; then
-	ad_subdomain_prefix="${given_ad_subdomainname}"
-fi
-
-# Determine Gluster volume names
-for volume in engine vmstore iso ctdb unixshare winshare blockshare backup; do
-	given_volume_name=$(sed -n -e "s/^.*hvp_${volume}_volumename=\\(\\S*\\).*\$/\\1/p" /proc/cmdline)
-	if [ -n "${given_volume_name}" ]; then
-		gluster_vol_name["${volume}"]="${given_volume_name}"
-	fi
-done
 
 # Determine network segments parameters
 fixed_mgmt_bondmode="false"
@@ -1210,6 +1189,8 @@ timezone ${local_timezone} --isUtc
 EOF
 
 # Create disk setup fragment
+in_use_devices=$(mount | awk '/^\/dev/ {print gensub("/dev/","","g",$1)}')
+kickstart_device=$(echo "${ks_dev}" | sed -e 's%^/dev/%%')
 # Note: all sizes in MiB except for data_block_size which is in KiB
 data_dev_size=$((${disk_device_size} / 1024 / 1024))
 data_block_size="2048"
@@ -1225,7 +1206,7 @@ clearpart --all --initlabel --disklabel=gpt
 # Bootloader placed on MBR, with 3 seconds waiting, with password protection, disabling high res text console, disabling CPU C-states and with I/O scheduler optimized for a virtualization/storage server
 bootloader --location=mbr --timeout=3 --password=${root_password} --append="nomodeset processor.max_cstate=1 intel_idle.max_cstate=0"
 # Ignore further disks
-# Note: further disks will be used as bricks later on
+# Note: some other disks will be used as bricks later on
 ignoredisk --only-use=${disk_device_name}
 # Automatically create UEFI or BIOS boot partition depending on hardware capabilities
 reqpart --add-boot
@@ -1261,11 +1242,55 @@ done
 if cat /sys/class/dmi/id/sys_vendor | egrep -q -v "(Microsoft|VMware|innotek|Parallels|Red.*Hat|oVirt|Xen)" ; then
 	# Note: resetting all disk devices since leftover configurations may interfer with installation and/or setup later on
 	for current_device in ${all_devices}; do
+		# Skipping devices in active use
+		if [ "${current_device}" = "${kickstart_device}" ] || echo "${in_use_devices}" | grep -q -w "${current_device}" ; then
+			continue
+		fi
 		dd if=/dev/zero of=/dev/${current_device} bs=1M count=10
 		dd if=/dev/zero of=/dev/${current_device} bs=1M count=10 seek=$(($(blockdev --getsize64 /dev/${current_device}) / (1024 * 1024) - 10))
 	done
 	partprobe
 	udevadm settle --timeout=5
+fi
+
+# Create install source selection fragment
+# Note: we use a non-local (hd:) stage2 location as indicator of network boot
+given_stage2=$(sed -n -e 's/^.*inst\.stage2=\(\S*\).*$/\1/p' /proc/cmdline)
+if echo "${given_stage2}" | grep -q '^hd:' ; then
+	# Detect use of NetInstall media
+	if [ -d /run/install/repo/repodata ]; then
+		# Note: we know that the local stage2 comes from a full DVD image (Packages repo included)
+		cat <<- EOF > /tmp/full-installsource
+		# Use the inserted optical media as in:
+		cdrom
+		# alternatively specify a NFS network share as in:
+		# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
+		# or an HTTP/FTP area as in:
+		#url --url https://dangerous.ovirt.life/hvp-repos/el7/os
+		EOF
+	else
+		# Note: since we detected use of NetInstall media (no local repo) we use network install source from CentOS mirrors
+		given_stage2="http://mirror.centos.org/centos/7/os/x86_64"
+		cat <<- EOF > /tmp/full-installsource
+		# Specify a NFS network share as in:
+		# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
+		# or an HTTP/FTP area as in:
+		url --url ${given_stage2}
+		# alternatively use the inserted optical media as in:
+		#cdrom
+		EOF
+	fi
+else
+	# Note: we assume that a remote stage2 has been copied together with the full media content preserving the default DVD structure
+	# TODO: we assume a HTTP/FTP area - add support for NFS
+	cat <<- EOF > /tmp/full-installsource
+	# Specify a NFS network share as in:
+	# nfs --opts=nolock --server NfsFqdnServerName --dir /path/to/CentOS/base/dir/copied/from/DVD/media
+	# or an HTTP/FTP area as in:
+	url --url ${given_stage2}
+	# alternatively use the inserted optical media as in:
+	#cdrom
+	EOF
 fi
 
 # Create bind zone files if running on master node
@@ -1891,7 +1916,7 @@ done
 %post --log /dev/console
 ( # Run the entire post section as a subshell for logging purposes.
 
-script_version="2018092902"
+script_version="2018121001"
 
 # Report kickstart version for reference purposes
 logger -s -p "local7.info" -t "kickstart-post" "Kickstarting for $(cat /etc/system-release) - version ${script_version}"
@@ -1994,7 +2019,7 @@ if grep -w -q 'hvp_nicmacfix' /proc/cmdline ; then
 	nicmacfix="true"
 fi
 
-# Determine choice of skipping custom packages intallation
+# Determine choice of skipping custom packages installation
 if grep -w -q 'hvp_orthodox' /proc/cmdline ; then
 	orthodox_mode="true"
 fi
@@ -2039,7 +2064,7 @@ fi
 mp=$(grep -w "/" /etc/fstab | sed -e 's/ .*//')
 if echo "$mp" | grep -q "^UUID="
 then
-    uuid=$(echo "$mp" |sed 's/UUID=//')
+    uuid=$(echo "$mp" | sed -e 's/UUID=//')
     rootdisk=$(blkid -U $uuid)
 elif echo "$mp" | grep -q "^/dev/"
 then
@@ -2057,7 +2082,8 @@ yum --enablerepo '*' clean all
 yum -y install yum-plugin-priorities
 
 # Add support for CentOS CR repository (to allow up-to-date upgrade later)
-yum-config-manager --enable cr > /dev/null
+# Note: a partially populated CR repo may introduce dependency-related errors - better to leave this to post-installation manual choices
+#yum-config-manager --enable cr > /dev/null
 
 # Add HVP custom repo
 yum -y --nogpgcheck install https://dangerous.ovirt.life/hvp-repos/el7/hvp/x86_64/hvp-release-7-5.noarch.rpm
@@ -2086,24 +2112,29 @@ if [ "${ovirt_release_package_suffix}" = "master" ]; then
 fi
 yum -y install http://resources.ovirt.org/pub/yum-repo/ovirt-release${ovirt_release_package_suffix}.rpm
 # If explicitly allowed, make sure that we use oVirt snapshot/nightly repos
-# Note: when nightly mode gets enabled we assume that we are late in the selected-oVirt lifecycle and some repositories and release packages may have disappeared - working around here
+# Note: when nightly mode gets enabled we assume that we are late in the selected-oVirt-version lifecycle and some repositories and release packages may have disappeared - working around here
 if [ "${ovirt_nightly_mode}" = "true" ]; then
-	# Disable CentOS-hosted oVirt and GlusterFS repositories
-	yum-config-manager --disable 'ovirt-centos-ovirt*' > /dev/null
-	yum-config-manager --disable 'ovirt-*-centos-gluster*' > /dev/null
+	# Marking CentOS-hosted repositories as skip_if_unavailable
+	for repo_name in $(grep -o '\[ovirt-.*centos.*\]' /etc/yum.repos.d/ovirt-*-dependencies.repo  | tr -d '[]'); do
+		yum-config-manager --save --setopt="${repo_name}.skip_if_unavailable=1" > /dev/null
+	done
 	# Manually add snapshot repositories
+	# Note: adding these manually since the release package may have disappeared
+	# Note: adding skip_if_unavailable since the repos too seem to disappear after a while
 	cat <<- EOF > "/etc/yum.repos.d/ovirt-${ovirt_version}-snapshot.repo"
 	[ovirt-${ovirt_version}-snapshot]
 	name=oVirt ${ovirt_version} - Nightly snapshot
 	baseurl=https://resources.ovirt.org/pub/ovirt-${ovirt_version}-snapshot/rpm/el\$releasever/
 	gpgcheck=0
 	enabled=1
+	skip_if_unavailable=1
 
 	[ovirt-${ovirt_version}-snapshot-static]
 	name=oVirt ${ovirt_version} - Nightly snapshot static
 	baseurl=https://resources.ovirt.org/pub/ovirt-${ovirt_version}-snapshot-static/rpm/el\$releasever/
 	gpgcheck=0
 	enabled=1
+	skip_if_unavailable=1
 	EOF
 	chmod 644 "/etc/yum.repos.d/ovirt-${ovirt_version}-snapshot.repo"
 fi
@@ -2133,6 +2164,8 @@ yum -y upgrade
 # TODO: Remove when fixed upstream
 # TODO: the following works only with re-instated CentOS6 fix above
 grubby --set-default=/boot/vmlinuz-$(rpm -q --last kernel | head -1 | cut -f 1 -d ' ' | sed -e 's/kernel-//')
+
+# Note: disabling includes in all yum commandline invocations to work around includepkgs lines inside oVirt dependency repo file - they are meant to avoid accidental upgrades according to https://bugzilla.redhat.com/show_bug.cgi?id=1552929
 
 # Install HAVEGEd
 # Note: even in presence of an actual hardware random number generator (managed by rngd) we install haveged as a safety measure
@@ -2266,6 +2299,11 @@ else
 	grub2_cfg_file="/etc/grub2.cfg"
 fi
 
+# Configure GRUB2 boot loader (no splash screen, no Plymouth, show menu, wait 5 seconds for manual override)
+# Note: alternatively, Plymouth may be instructed to use detailed listing with: plymouth-set-default-theme -R details
+sed -i -e '/^GRUB_CMDLINE_LINUX/s/\s*rhgb//' -e '/^GRUB_TIMEOUT/s/=.*$/="5"/' /etc/default/grub
+grub2-mkconfig -o "${grub2_cfg_file}"
+
 # Setup a serial terminal
 serial_found="false"
 for link in /sys/class/tty/*/device/driver ; do
@@ -2284,11 +2322,6 @@ if [ "${serial_found}" = "true" ]; then
 	EOF
 	grub2-mkconfig -o "${grub2_cfg_file}"
 fi
-
-# Configure GRUB2 boot loader (no splash screen, no Plymouth, show menu, wait 5 seconds for manual override)
-# Note: alternatively, Plymouth may be instructed to use detailed listing with: plymouth-set-default-theme -R details
-sed -i -e '/^GRUB_CMDLINE_LINUX/s/\s*rhgb//' -e '/^GRUB_TIMEOUT/s/=.*$/="5"/' /etc/default/grub
-grub2-mkconfig -o "${grub2_cfg_file}"
 
 # Conditionally add memory test entry to boot loader
 if dmidecode -s system-manufacturer | egrep -q -v "(Microsoft|VMware|innotek|Parallels|Red.*Hat|oVirt|Xen)" ; then
@@ -3672,9 +3705,11 @@ fi
 
 # TODO: perform NetworkManager workaround configuration on interfaces as detected in pre section above - remove when fixed upstream
 for file in /tmp/hvp-networkmanager-conf/ifcfg-* ; do
-	cfg_file_name=$(basename ${file})
-	sed -i -e '/^DEFROUTE=/d' -e '/^MTU=/d' ${ANA_INSTALL_PATH}/etc/sysconfig/network-scripts/${cfg_file_name}
-	cat ${file} >> ${ANA_INSTALL_PATH}/etc/sysconfig/network-scripts/${cfg_file_name}
+	if [ -f "${file}" ]; then
+		cfg_file_name=$(basename "${file}")
+		sed -i -e '/^DEFROUTE=/d' -e '/^MTU=/d' "${ANA_INSTALL_PATH}/etc/sysconfig/network-scripts/${cfg_file_name}"
+		cat "${file}" >> "${ANA_INSTALL_PATH}/etc/sysconfig/network-scripts/${cfg_file_name}"
+	fi
 done
 
 # Save exact pre-stage environment
